pyproject.toml
[project]
name = "pixel-art-finetune"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.1.1",
    "groq>=0.31.1",
    "huggingface-hub>=0.35.0",
    "numpy>=2.2.6",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
]
